Fix 9 MEDIUM serial/LIN bugs from deep code review
- UART: 9-bit TX now combines TXB8 (UCSRB bit 0) with UDR as u16 shift reg
- UART: 9-bit RX stores 9th bit in UCSRB RXB8; inject/consume use u16
- SPI: SPIF cleared on SPDR write in master mode when starting a new transfer
- SPI8X: CPOL SCK polarity now checked at correct bit 2 (0x04), not bit 4 (0x10)
- UART8X: 9-bit RX passes bit9 to FIFO instead of always false
- UART8X: RXDATAH now latches on RXDATAL read (correct hardware behavior)
- EUSART: consume_transmitted_byte returns data only after TX completes
- LIN: checksum verification uses additive complement (0xFF - sum) not bitwise NOT
- LIN: LIDST field shifted by 4 (bits 4-9) not 5 (bits 5-10)

221/221 tests pass.

+++ include/vioavr/core/uart.hpp
@@ -30,8 +30,8 @@ public:
 
     [[nodiscard]] ClockDomain clock_domain() const noexcept override;
 
-    void inject_received_byte(u8 data) noexcept;
-    bool consume_transmitted_byte(u8& data) noexcept;
+    void inject_received_byte(u16 data) noexcept;
+    bool consume_transmitted_byte(u16& data) noexcept;
 
 private:
     std::string name_;
@@ -61,7 +61,7 @@ private:
     u64 rx_cycle_accumulator_ {};
     bool tx_buffer_full_ {};
 
-    std::deque<u8> tx_output_queue_;
+    std::deque<u16> tx_output_queue_;
 
     void update_pin_ownership() noexcept;
     [[nodiscard]] bool power_reduction_enabled() const noexcept;
